requirements.txt
openai-whisper
ffmpeg-python
torch
yt-dlp
sentencepiece
pinecone[grpc]
sentence-transformers
flask